requirements.txt
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
opencv-python==4.9.0.80
scikit-image==0.22.0
rasterio==1.3.9
tifffile==2024.2.12
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
#Framework ML (ajouté à l'étape 4 selon ton choix) :
tensorflow-cpu==2.16.1
torch==2.3.1
torchvision==0.18.1